CSc8830_Module6_Assignment/requirements.txt
# Core dependencies
openai>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
open-clip-torch>=2.20.0
pygame>=2.5.0
pyyaml>=6.0.0
